update all classes

+++ src/HoursReportUtility/src/ExcelWriter.java
@@ -12,6 +12,7 @@ import javax.swing.DefaultListModel;
 import javax.swing.JFileChooser;
 import javax.swing.JFrame;
 import javax.swing.JList;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTextArea;
@@ -30,44 +31,6 @@ import org.openxmlformats.schemas.spreadsheetml.x2006.main.CTFilterColumn;
 public class ExcelWriter {
 	static JFileChooser fileChooser;
 	public static final String OUTPUT_FILE = "Hours_" + FileUtility.getTimeStamp() + ".xlsx";
-
-	public void filterGUI(List<Cell> list) {
-		DefaultListModel<String> listModel = new DefaultListModel<>();
-		FileWriter writer = new FileWriter();
-		writer.removeDupes(list);
-		for (int i = 0; i < list.size(); i++) {
-			listModel.addElement(list.get(i).getStringCellValue());
-		}
-		JList<String> employeeList = new JList(listModel);
-		JFrame frame = new JFrame();
-		JTextArea textArea = new JTextArea(15, 20);
-		textArea.setEditable(false);
-		JPanel topPane = new JPanel();
-		JPanel bottomPane = new JPanel();
-		topPane.add(new JScrollPane(employeeList));
-		bottomPane.add(textArea);
-		frame.add(bottomPane, BorderLayout.PAGE_END);
-		frame.add(topPane, BorderLayout.PAGE_START);
-		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		frame.setTitle("Employee Filter");
-		frame.setSize(200, 200);
-		frame.setLocationRelativeTo(null);
-		frame.pack();
-		frame.setVisible(true);
-		
-		employeeList.setSelectionMode(ListSelectionModel.MULTIPLE_INTERVAL_SELECTION);
-		employeeList.addListSelectionListener((new ListSelectionListener() {
-			@Override
-			public void valueChanged(ListSelectionEvent e) {
-				if (!e.getValueIsAdjusting()) {
-					String employee = employeeList.getSelectedValue();
-					textArea.append(employee + "\n");
-				}
-			}
-		}));
-
-		
-	}
 
 	public static void outputFile(List<Cell> projects, List<Cell> dates, List<Cell> names, List<Cell> duration,
 			List<Cell> status) throws IOException {
@@ -126,12 +89,23 @@ public class ExcelWriter {
 
 		}
 		sheet0.autoSizeColumn(0);
-		sheet0.setAutoFilter(CellRangeAddress.valueOf("B1"));
-		ExcelWriter writer = new ExcelWriter();
-		writer.filterGUI(names);
 
-		FileOutputStream out = new FileOutputStream(OUTPUT_FILE);
-		wb.write(out);
+		int result = JOptionPane.showConfirmDialog(null, "Filter Results?");
+
+		if (result == JOptionPane.OK_OPTION) {
+			FileOutputStream out = new FileOutputStream(OUTPUT_FILE);
+			wb.write(out);
+			wb.close();
+			Filter filter = new Filter();
+
+			// if OK the filter GUI will appear and user can select people
+			filter.filterGUI();
+		} else if (result == JOptionPane.NO_OPTION || result == JOptionPane.CANCEL_OPTION) {
+			// if no then the output is written to the output file
+			FileOutputStream out = new FileOutputStream(OUTPUT_FILE);
+			wb.write(out);
+			wb.close();
+		}
 
 	}
